chore: rename checkpoint_manager to pytorch_checkpoint_manager
Also rename old pytorch_checkpoint_manager to filesystem checkpoint manager
